% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{wordpiece_vocab}
\alias{wordpiece_vocab}
\title{Load a wordpiece Vocabulary}
\usage{
wordpiece_vocab()
}
\value{
A wordpiece_vocabulary.
}
\description{
A wordpiece vocabulary is a named integer vector with class
"wordpiece_vocabulary". The names of the vector are the tokens, and
the values are the integer identifiers of those tokens. The vocabulary is
0-indexed for compatibility with Python implementations.
}
\examples{
head(wordpiece_vocab())
}
